Use more sensible name for quotas
And also change to crate collector to no longer have full control over the score of the game.

+++ Project/Assets/_Data/Scripts/Crates/CrateExtensions.cs
@@ -21,7 +21,7 @@ public static class CrateExtensions
     public struct ScheduleQuota
     {
         public CrateTag requiredTag;
-        public int requiredScore;
+        public float requiredScore;
         [Min(0f)]
         public float timeLimit;
     }
